+++ crates/apps/src/lib/node/ledger/shell/finalize_block.rs
@@ -82,7 +82,8 @@ where
             self.state.in_mem().update_epoch_blocks_delay
         );
 
-        // Finalize the transactions' hashes from the previous block
+        // Finalize the transactions' hashes from the previous block. Also cache
+        // "last" hashes from the previous block in case of a rollback
         let (write_log, _in_mem, db) = self.state.split_borrow();
         for (raw_key, _, _) in db.iter_replay_protection() {
             let hash = raw_key.parse().expect("Failed hash conversion");
@@ -114,18 +115,18 @@ where
             req.byzantine_validators,
         )?;
 
-        // Take IBC events that may be emitted from PGF
-        for ibc_event in self.state.write_log_mut().take_ibc_events() {
-            let mut event = Event::from(ibc_event.clone());
-            // Add the height for IBC event query
-            let height = self.state.in_mem().get_last_block_height() + 1;
-            event["height"] = height.to_string();
-            response.events.push(event);
-        }
-
         if new_epoch {
             // Apply PoS and PGF inflation
             self.apply_inflation(current_epoch)?;
+
+            // Take IBC events that may be emitted from PGF
+            for ibc_event in self.state.write_log_mut().take_ibc_events() {
+                let mut event = Event::from(ibc_event.clone());
+                // Add the height for IBC event query
+                let height = self.state.in_mem().get_last_block_height() + 1;
+                event["height"] = height.to_string();
+                response.events.push(event);
+            }
         }
 
         let mut stats = InternalStats::default();