+++ crates/apps/src/lib/node/ledger/shell/init_chain.rs
@@ -1,5 +1,5 @@
 //! Implementation of chain initialization for the Shell
-use std::collections::HashMap;
+use std::collections::{BTreeMap, HashMap};
 use std::ops::ControlFlow;
 
 use masp_primitives::merkle_tree::CommitmentTree;
@@ -12,6 +12,7 @@ use namada::ledger::parameters::Parameters;
 use namada::ledger::{ibc, pos};
 use namada::proof_of_stake::BecomeValidator;
 use namada::state::StorageWrite;
+use namada::token::storage_key::masp_token_map_key;
 use namada::token::{credit_tokens, write_denom};
 use namada::vm::validate_untrusted_wasm;
 use namada_sdk::eth_bridge::EthBridgeStatus;
@@ -27,6 +28,7 @@ use crate::config::genesis::transactions::{
     BondTx, EstablishedAccountTx, Signed as SignedTx, ValidatorAccountTx,
 };
 use crate::facade::tendermint_proto::google::protobuf;
+use crate::node::ledger;
 use crate::wasm_loader;
 
 /// Errors that represent panics in normal flow but get demoted to errors
@@ -93,6 +95,38 @@ where
                 "Current chain ID: {}, Tendermint chain ID: {}",
                 chain_id, init.chain_id
             )));
+        }
+        if ledger::migrating_state().is_some() {
+            let rsp = response::InitChain {
+                validators: self
+                    .get_abci_validator_updates(true, |pk, power| {
+                        let pub_key: crate::facade::tendermint::PublicKey =
+                            pk.into();
+                        let power =
+                            crate::facade::tendermint::vote::Power::try_from(
+                                power,
+                            )
+                            .unwrap();
+                        validator::Update { pub_key, power }
+                    })
+                    .expect("Must be able to set genesis validator set"),
+                app_hash: self
+                    .state
+                    .in_mem()
+                    .merkle_root()
+                    .0
+                    .to_vec()
+                    .try_into()
+                    .expect("Infallible"),
+                ..Default::default()
+            };
+            debug_assert!(!rsp.validators.is_empty());
+            debug_assert!(
+                !Vec::<u8>::from(rsp.app_hash.clone())
+                    .iter()
+                    .all(|&b| b == 0)
+            );
+            return Ok(rsp);
         }
 
         // Read the genesis files
@@ -418,6 +452,7 @@ where
 
     /// Init genesis token accounts
     fn init_token_accounts(&mut self, genesis: &genesis::chain::Finalized) {
+        let mut token_map = BTreeMap::new();
         for (alias, token) in &genesis.tokens.token {
             tracing::debug!("Initializing token {alias}");
 
@@ -438,13 +473,12 @@ where
                 // add token addresses to the masp reward conversions lookup
                 // table.
                 let alias = alias.to_string();
-                self.state
-                    .in_mem_mut()
-                    .conversion_state
-                    .tokens
-                    .insert(alias, address.clone());
+                token_map.insert(alias, address.clone());
             }
         }
+        self.state
+            .write(&masp_token_map_key(), token_map)
+            .expect("Couldn't init token accounts");
     }
 
     /// Init genesis token balances
@@ -937,7 +971,6 @@ impl<T> Policy<T> {
 
 #[cfg(all(test, not(feature = "integration")))]
 mod test {
-    use std::collections::BTreeMap;
     use std::str::FromStr;
 
     use namada::core::string_encoding::StringEncoded;